requirements.txt
addict==2.4.0
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.16.1
appdirs==1.4.4
brotlipy==0.7.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.7
colorama==0.4.6
contourpy==1.1.0
crcmod==1.7
cryptography==41.0.2
cycler==0.11.0
docker-pycreds==0.4.0
einops==0.6.1
filelock==3.9.0
fonttools==4.42.1
gitdb==4.0.10
gitpython==3.1.32
gmpy2==2.1.2
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
jinja2==3.1.2
jmespath==0.10.0
kiwisolver==1.4.5
markdown-it-py==3.0.0
markdown==3.4.4
markupsafe==2.1.1
mat4py==0.5.0
matplotlib==3.7.2
mdurl==0.1.2
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
mmcv==2.0.1
mmdet==3.1.0
mmengine==0.8.4
mmpretrain==1.0.2
mmyolo==0.6.0
model-index==0.1.11
modelindex==0.0.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
opencv-python==4.8.0.76
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.21
ordered-set==4.1.0
oss2==2.17.0
packaging==23.1
pandas==2.0.3
pathtools==0.1.2
pillow==9.4.0
pip==23.2.1
platformdirs==3.10.0
prettytable==3.8.0
protobuf==4.24.2
psutil==5.9.5
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.18.0
pygments==2.16.1
pyopenssl==23.2.0
pyparsing==3.0.9
pysocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
pyyaml==6.0.1
requests==2.28.2
rich==13.4.2
scipy==1.10.1
sentry-sdk==1.29.2
setproctitle==1.3.2
setuptools==60.2.0
shapely==2.0.1
six==1.16.0
smmap==5.0.0
sympy==1.11.1
tabulate==0.9.0
termcolor==2.3.0
terminaltables==3.1.10
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.2
triton==2.0.0
typing-extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
wandb==0.15.8
wcwidth==0.2.6
wheel==0.38.4
yapf==0.40.1
zipp==3.16.2